Added Billing Cycles to explain the stream lifecycles
Now the VirtualCard is responsible (and the source of truth) for opening new BillingCycle, and BillingCycle accordingly for the clousure.

Streams are integrated through events. Currently that happens through subscriptions and event handler, but that could be easily made synchronous if event handler was called from the billing service explicitly.